improve handling of translator modules
