+++ src/translators/python/impure/pip/generate-generic-lock.py
